add irrigation strategies
