+++ src/components/CropRotation/index.tsx
@@ -2,25 +2,27 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import Box from '@mui/material/Box';
-import { CircularProgress, Typography } from '@mui/material';
 
-import { useSelectedResearch } from '@app/utils/hooks';
-import { theme } from '@app/theme';
+import { useSelectedResearch, useCropRotationYieldData } from '@app/utils/hooks';
 import { DataStateContext } from '@app/store/contexts';
 import Header from '../childComponents/navigation/Header';
 
-import ResearchLeftSidebar from '../childComponents/ResearchLeftSidebar';
+import ResearchLeftSidebar from '@app/components/childComponents/ResearchLeftSidebar';
 import CropRotationYield from './CropRotationYield';
+import withLoading from '@app/components/childComponents/hocs/withLoading';
+import withErrorHandling from '@app/components/childComponents/hocs/withErrorHandling';
 
-const CropRotation = (): JSX.Element => {
-    const { research_id } = useParams<{ research_id: string }>();
+const CropRotationYieldWithLoading = withLoading(CropRotationYield);
+const CropRotationYieldWithErrorHandling = withErrorHandling(CropRotationYieldWithLoading);
+
+const CropRotationComponent: React.FC<{ research: ResearchDetail | null; research_id: string | undefined }> = ({
+    research,
+    research_id
+}): JSX.Element => {
     const { selectedResearch } = React.useContext(DataStateContext);
-    let loading = false;
-    let error = null;
-    let research = null;
-    if (!selectedResearch && research_id) {
-        [research, loading, error] = useSelectedResearch(research_id);
-    }
+    const [cropRotationYieldData, cropRotationYieldDataLoading, cropRotationYieldDataError] = useCropRotationYieldData(
+        selectedResearch ? selectedResearch.id : research_id
+    );
     const leftSidebarDetails = {
         dataType: 'Crop Rotation',
         pi: '-',
@@ -28,39 +30,43 @@ const CropRotation = (): JSX.Element => {
         introduction: '-',
         conclusion: '-'
     };
+
+    return (
+        <Box sx={{ display: 'flex' }}>
+            <ResearchLeftSidebar {...{ selectedResearch: selectedResearch ?? research }} {...leftSidebarDetails} />
+            <CropRotationYieldWithErrorHandling
+                error={cropRotationYieldDataError}
+                isLoading={cropRotationYieldDataLoading}
+                cropRotationYieldData={cropRotationYieldData}
+            />
+        </Box>
+    );
+};
+
+const CropRotationComponentWithLoading = withLoading(CropRotationComponent);
+const CropRotationComponentWithErrorHandling = withErrorHandling(CropRotationComponentWithLoading);
+
+const CropRotation = (): JSX.Element => {
+    const { research_id } = useParams<{ research_id: string }>();
+    const { selectedResearch } = React.useContext(DataStateContext);
+    let loading = false;
+    let error: string | null = null;
+    let research: ResearchDetail | null = null;
+    if (!selectedResearch && research_id) {
+        [research, loading, error] = useSelectedResearch(research_id);
+    }
+
     return (
         <Box>
             <Box sx={{ pointerEvents: 'auto' }}>
                 <Header />
             </Box>
-            {loading ? (
-                <Box display="flex" justifyContent="center" justifyItems="center" sx={{ height: '100vh' }}>
-                    <CircularProgress />
-                </Box>
-            ) : error === null ? (
-                <Box sx={{ display: 'flex' }}>
-                    <ResearchLeftSidebar
-                        {...{ selectedResearch: selectedResearch ?? research }}
-                        {...leftSidebarDetails}
-                    />
-                    <CropRotationYield />
-                </Box>
-            ) : (
-                <Typography
-                    variant="h6"
-                    sx={{
-                        font: 'Poppins',
-                        fontWeight: 400,
-                        fontSize: '16px',
-                        lineHeight: '25.6px',
-                        letterSpacing: '0.15px',
-                        marginRight: '5px',
-                        color: theme.palette.text.primary
-                    }}
-                >
-                    {error}
-                </Typography>
-            )}
+            <CropRotationComponentWithErrorHandling
+                research={selectedResearch ?? research}
+                research_id={research_id}
+                error={error}
+                isLoading={loading}
+            />
         </Box>
     );
 };
