+++ src/components/IrrigationStrategies/index.tsx
@@ -2,15 +2,41 @@ import React from 'react';
 import { useParams } from 'react-router-dom';
 
 import Box from '@mui/material/Box';
-import { CircularProgress, Typography } from '@mui/material';
 
 import { useSelectedResearch } from '@app/utils/hooks';
-import { theme } from '@app/theme';
 import { DataStateContext } from '@app/store/contexts';
-import Header from '../childComponents/navigation/Header';
+import Header from '@app/components/childComponents/navigation/Header';
 
-import ResearchLeftSidebar from '../childComponents/ResearchLeftSidebar';
+import ResearchLeftSidebar from '@app/components/childComponents/ResearchLeftSidebar';
 import IrrigationStrategiesPage from './IrrigationStrategiesPage';
+import withLoading from '@app/components/childComponents/hocs/withLoading';
+import withErrorHandling from '@app/components/childComponents/hocs/withErrorHandling';
+
+const IrrigationStrategiesComponent: React.FC<{ research: ResearchDetail | null; research_id: string | undefined }> = ({
+    research
+    // research_id
+}): JSX.Element => {
+    const { selectedResearch } = React.useContext(DataStateContext);
+
+    const leftSidebarDetails = {
+        dataType: 'Irrigation Strategies',
+        pi: '-',
+        contactInfo: '-',
+        introduction:
+            'This is an introduction to the irrigation strategies research. It will be updated with more information soon.',
+        conclusion: 'Seed 1 is more drought tolerant than seed 2'
+    };
+
+    return (
+        <Box sx={{ display: 'flex' }}>
+            <ResearchLeftSidebar {...{ selectedResearch: selectedResearch ?? research }} {...leftSidebarDetails} />
+            <IrrigationStrategiesPage />
+        </Box>
+    );
+};
+
+const IrrigationStrategiesComponentWithLoading = withLoading(IrrigationStrategiesComponent);
+const IrrigationStrategiesComponentWithErrorHandling = withErrorHandling(IrrigationStrategiesComponentWithLoading);
 
 const IrrigationStrategies = (): JSX.Element => {
     const { research_id } = useParams<{ research_id: string }>();
@@ -22,48 +48,18 @@ const IrrigationStrategies = (): JSX.Element => {
     if (!selectedResearch && research_id) {
         [research, loading, error] = useSelectedResearch(research_id);
     }
-    const leftSidebarDetails = {
-        dataType: 'Irrigation Strategies',
-        pi: '-',
-        contactInfo: '-',
-        introduction:
-            'This is an introduction to the irrigation strategies research. It will be updated with more information soon.',
-        conclusion: 'Seed 1 is more drought tolerant than seed 2'
-    };
 
     return (
         <Box>
             <Box sx={{ pointerEvents: 'auto' }}>
                 <Header />
             </Box>
-            {loading ? (
-                <Box display="flex" justifyContent="center" justifyItems="center" sx={{ height: '100vh' }}>
-                    <CircularProgress />
-                </Box>
-            ) : error === null ? (
-                <Box sx={{ display: 'flex' }}>
-                    <ResearchLeftSidebar
-                        {...{ selectedResearch: selectedResearch ?? research }}
-                        {...leftSidebarDetails}
-                    />
-                    <IrrigationStrategiesPage />
-                </Box>
-            ) : (
-                <Typography
-                    variant="h6"
-                    sx={{
-                        font: 'Poppins',
-                        fontWeight: 400,
-                        fontSize: '16px',
-                        lineHeight: '25.6px',
-                        letterSpacing: '0.15px',
-                        marginRight: '5px',
-                        color: theme.palette.text.primary
-                    }}
-                >
-                    {error}
-                </Typography>
-            )}
+            <IrrigationStrategiesComponentWithErrorHandling
+                research={selectedResearch ?? research}
+                research_id={research_id}
+                error={error}
+                isLoading={loading}
+            />
         </Box>
     );
 };
